requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docx==0.2.4
docx2pdf==0.1.8
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.103.2
frozenlist==1.3.3
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==2.11.3
lxml==5.1.0
magic-filter==1.0.12
Mako==1.2.4
MarkupSafe==2.1.5
multidict==6.0.5
openpyxl==3.0.10
packaging==21.3
Pillow==9.5.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pyngrok==7.0.5
pyparsing==3.1.2
pypng==0.20220715.0
python-docx==1.1.0
python-dotenv==0.21.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlmodel==0.0.16
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.7.1
uvicorn==0.22.0
yarl==1.9.4
zipp==3.15.0
aiogram~=2.25.2
aiocache~=0.12.3
setuptools~=68.0.0